upd
Note: It is decided not to use bevy for now since the full engine like
bevy is an overkill for game I want to make at start but bevy is under
strong consideration for the future projects. Now raylib-rs has been
chosen. Why not macroquad? Macroquad did not provide some functionality
I needed.